Tweaks to make codacy happy
